* Modules/Preferences/Build/PCBuildPrefs.m         * Framework/PCButton.m         * Modules/Parsers/ProjectCenter/ObjCMethodHandler.m           shutup some clang warnings         * Framework/PCProjectBrowser.m           NSInteger conversions
OK Fred 



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/projectcenter/trunk@36098 72102866-910b-0410-8b05-ffd578937521

+++ Framework/PCProjectBrowser.m
@@ -308,11 +308,11 @@ NSString *PCBrowserDidSetPathNotification = @"PCBrowserDidSetPathNotification";
 
 - (void)reloadLastColumnAndNotify:(BOOL)yn
 {
-  int	    column = [browser lastColumn];
+  NSInteger column = [browser lastColumn];
   NSString  *category = [self nameOfSelectedCategory];
-  int       selectedColumn = [browser selectedColumn];
+  NSInteger selectedColumn = [browser selectedColumn];
   NSMatrix  *colMatrix = [browser matrixInColumn:selectedColumn];
-  int       rowCount = 0, colCount = 0, spCount = 0;
+  NSInteger rowCount = 0, colCount = 0, spCount = 0;
   PCProject *activeProject = [[project projectManager] activeProject];
   NSString  *selCellTitle = [[browser selectedCell] stringValue];
